Add fixed argument to variadic APIs, librebol.h cleanup
In order to know what environment to do binding in for API calls, it's
needed to know what stack level we are in (e.g. which native or module).
But this is not knowable beyond a certain moment of entry to a native's
code, because it could call other helper routines that are not intended
to see things like that function's arguments.

This lays the groundwork for being able to capture a named variable at
callsites which would exist globally in C files but then be overridden
by a redefinition of that name inside the native.  By making the API
macros automatically pick up this and pass it through, it should be
possible to eliminate things like rebArgR() entirely.

As it's only the first step, this just passes a nullptr through to be
the specifier that the variadic receives.  But it does a lot of other
cleanup in the process of getting that to work, so that the generated
rebol.h is easier to comprehend and debug.

This also simplifies things because variadic macros will be a fundamental
feature required, so pure C89 won't be possible any longer (though some
compilers implemented variadic macros before that.)