Add new comma and space behavior to `$<>` in CSCAPE
This adds the ability to say $<foo,> and it will delimit foo with comma.
If foo is an empty block it will vanish (and potentially the whole line
will vanish).

If you say $<xxx > then if the contents don't vanish, a space will be
after them.  But if it vanishes, there will be no space.

These can be combined, so that $<xxx, > will delimit with ", "