Auto include librebol/sys-core for extensions
The desire to be able to coordinate what LIBREBOL_SPECIFIER is in
extensions (in order to pick up the specifier inside functions or out)
means it's best if the #include "rebol.h" is done by the same file
that defines what DECLARE_NATIVE() means.

This does the necessary juggling to prepare for that, and also breaks
the dependency on %sys-core.h for the tmp-mod-init.c files when
USE-LIBREBOL is true in the config...meaning someone could compile
an extension without having the core on their machine at all (at least
in theory).