Some untangling of %sys-ext.h
The dependencies of %sys-ext.h were a bit unclear.  This moves the
builtin extensions table--as well as the API table--into their own
.c files so that what #includes they use can be just what's needed
for the table and not get mixed up with the file they're .inc'd into.

It also makes the ExtensionCollator definition part of %sys-core.h
since in general extensions shouldn't need to define that or the
IDX_ values.