Add shadowing specifier for API natives
This implements the concept that each native that uses the librebol API
interface will declare a local variable that can be picked up by the
macros of `rebValue()` etc.  This variable will be what is used as the
specifier for the evaluations, and it knows specifically about the frame
of the native it is running in.  This allows access to local variables
without using things like rebArg() or rebArgR().