Rename RL_rebXXX() functions to API_rebXXX() functions
They have to be named something different than rebXXX(), because what rebXXX()
is depends on whether the API is being called through a table or not, as well
as the need to run through variadic helper macros for adding rebEND and
capturing the LIBREBOL_SPECIFIER.

But RL_rebXXX() is an antiquated naming scheme that meant "(R)ebol (L)ibrary".
API is much clearer.