Narrow BLOCK!/GROUP! initialization to generic arrays
There are a lot of different "Array" types in the system, since they are
used to implement things like varlists...which can have antiforms and
voids in them as the value of variables.

This asserts you don't try to initialize a BLOCK! or GROUP! with any of
the arrays that have antiforms or voids legal.  Nothing triggers this
assert at the moment, but it's a good protection against that happening
in the future.