Kill off grievous @ hack which decayed quasiforms
Suppressing evaluation in the API was at first done with rebQ():

     rebElide("append block", rebQ(word));

A nice trick came along to use the @ operator, which was cleaner:

     rebElide("append block @", word);

But by design you can't splice antiforms in feeds.  So this wouldn't work for
antiforms like NULL/etc.

A hack was put in that the feed splicing mechanic would splice antiforms as
quasiforms, and then @ would turn quasiforms into antiforms:

     >> @ ~null~
     == ~null~  ; anti

It gave the impression of working, because it wasn't too common to use
quasiforms.  But it had negative effects of splicing antiforms in places that
were not expecting them (e.g. actions would wind up not running when they
should likely have run), and it was broken in handling quasiforms.

This redoes the mechanism so the feed splices a meta form for antiforms and
voids, with a note about the fact that it has done that.  Only the @ operator
will tolerate the situation, other code will error when it reaches that
point in the feed.