Mass Renaming: REBVAL => Value, RELVAL => Cell
In order to make copying code between R3C and current Ren-C easier
(and so it's readable to modern eyes without too much annoyance)
this kicks off the renaming of REBVAL to Value.  RELVAL doesn't
really have a good analogue since it doesn't exist in modern Ren-C,
but just calling it Cell at least makes it so one of the two
cell subclasses is named "Cell".