Rename string macros to reflect UCS-2 Nature
R3C will not be migrated to UTF-8 Everywhere.  It's too much work to
do again (!) and the UCS-2 code is simple and predictable enough for
bootstrapping purposes.

Rename things to make it clear what point in the migration was reached,
and that it's not merging Symbol and String in this branch.