Drop SECURE Dialect
SECURE was an unfinished feature, and the security it offered would not
be satisfactory for any real applications even had it been finished.

With the advent of WebAssembly and other container technologies, it's
best to leave the design of security to the platform, and focus on
solving the language itself.

(Even if SECURE were pursued in the main branch, it would not be grafted
onto R3C.)