Kill off VALUE_FLAG_UNEVALUATED
The unevaluated flag was always a kind of problematic idea, but it seemed
to have some benefits in helping newbies avoid `if [x > y] [...]` types of
mistakes.

However, with the new binding model everything's out the window, because
now blocks are "evaluative" in terms of capturing their environment.  The
feature was killed in mainline, so no need to have it complicating R3C.