Get rid of NODE_FLAG_STACK mechanics
There was a speculative idea that cells could encode what stack level
they lived on, and this could be exploited to avoid making entities
that needed to be garbage collected if the references couldn't
"escape" the blocking of the stack levels...so that when a function
was over it could just free a frame's varlist without managing it.

This idea was never implemented, and it's out the window now due to
the fact that the future is a stackless optimization.