Get rid of janky old object rendering concept
Prior to the existence of isotopes, there was no real answer of how things
like NULL that couldn't be put in blocks should be rendered in MAKE
OBJECT!.  So there was a weird idea of listing the field names in one
block (a "spec") and then if there wasn't an assignment in the second
block (the "body") the lack of assignment could cue you to the missing
value.

It's now understood that antiforms are meta-representable as quasiforms.
That's not implemented in this old branch of R3C, but we can still give
the rendering equivalent by showing NULL as ~null~.