Rework DO and LOAD to run script-relative with TAG!
It's desirable to make the command lines when running %make.r compatible with
modern conventions (hence being able to do things like use the same GitHub
workflow files).

In order to do that, the paths passed to make.r need to be relative to the
directory in which the interpreter is invoked, not relative to the path where
make.r lives.  This does the tweaking to get there by shimming the convention
of what DO of a TAG! means to be relative to system/script/path