Sync dtoa.c code with version from internet (circa 2001)
The file %f-dtoa.c is an adaptation of a widely distributed "Decimal
to ASCII" conversion code that originated from Lucent/AT&T.  It was
included in R3-Alpha to mold decimals--minus the `strtod()` function
implementation (which is part of the C runtime, and would cause a
duplicate symbol error at link time if included.)

There were several deviations in the dtoa() from the Internet code.  It
is not known what accounts for all of these deviations--however, the
Emscripten "emterpreter" build was failing somewhat catastrophically.
When investigating the question of if a bug against the emterpreter
should be filed, the first thing done was to try a test against the
published code from the Internet.  No failure was observed.

Given a lack of rationale or documentation for why the Rebol file
should not match the code from the internet, this sync'd it up with
the published version.  A later version exists that is a package which
is not distributed as a single file, which is dated 2006 instead of
2001...but integrating this would be a larger challenge (as building
it requires a third-party configuration step that the Rebol build
process wishes to avoid).