Eliminate Visual Studio Project Generation
The creation of Rebmake was an attempt to unify the divergent efforts
of three distinct build methods:

* A small makefile generator from R3-Alpha
* The CMake usage of Atronix
* A build.reb script based on using Rebol itself with CALL

The ambitious idea was to essentially write something higher level than
CMake, which could be used to generate makefiles, CMake files, Visual
Studio XML projects, or drive the build with CALL.

This vast scope was contentious with Rebol's goals of complexity
elimination, as it intrinsically wound up making a build system more
complicated than any of the things it was wrapping.

In the intervening time, VSCode has become the editor of choice for
most languages...so catering to Visual Studio no longer makes sense.
With NMAKE support bolstered a bit--now running on GitHub Actions--
this goes ahead and drops the VSCode project generation.