Make JOIN non-mutating, simplify usages
There was a strange thought that since all the other operations like
APPEND and such were mutating, that JOIN should be too--since its
semantics might be useful.

This turned out to not serve much purpose, and in fact JOIN's key
application is not as a variant of APPEND but rather something you
can use in cases where you can't append--e.g. to modern Ren-C's
immutable paths.

JOIN does not reduce in modern Ren-C, and in fact is used to do
things like append blocks to paths as a mechanical operation:

    >> join 'a/ [b c]
    == a/[b c]