Implement the TAG!-relative-DO/LOAD, don't change directory
The new concept is not to switch to the directory of a script when
running it, but to leave the directory where the user was.  This makes
interpretations of paths passed on the command line more sensible as
a default, and if you want to do something script-relative that is
the special case.

One convenience for doing script-relative things is to take script
paths in a tag, and have that mean relative to the current directory.

(This may someday be changed to just plain strings.)