Don't LOCK source on load
This R3C branch will never receive the CONST+MUTABLE feature, which is
the believed correctly implemented answer to "about the best it can
get" as far as convenience vs. safety for mutability of data.

Given that it won't, the LOCK of source is overly restrictive,
especially when trying to bend this interpreter for bootstrap.  Remove
the restriction.