VOID as ~void~ antiform, TRASH as BLANK! antiform
This simplifies checks for what's legal in blocks, by eliminating the
separate VOID class which wasn't an antiform yet still could not be
put in blocks at quote level zero.

Most client code should be unchanged, but it provides new solidity to
the REIFY and DEGRADE operations... as they only need to operate on
antiforms/quasiforms, and have full coverage including void for
being able to reify things.