Switch to ~null~ and ~void~ (vs `<opt>` and `<void>`)
A long time ago when there was only one state you could put in values
that could not be in blocks, it was called VOID.  It didn't have a
datatype, so some alternative signal was needed in the type spec
blocks.  The idea of using a TAG! and saying `<opt>` was chosen
as a temporary solution.

That temporary solution wound up lasting for many years, as the lack
of a satisfactory answer persisted.  The type changed from being VOID
to being NULL, and then something called VOID came along with a similar
problem.  So `<void>` was used for that.

As the needs to be more expressive about the properties of parameters
grew, tags began to control other features...making it a bit saturated
to also be using tags for void and null indications.  The modern idea
is to use quasiforms to represent their antiforms, and it seemed like
a pretty reasonable solution...but didn't fit the situation with void
not being an antiform of the word "void"

But that's changed, and now void *is* the antiform of the word void.
So ~void~ makes a good indicator in the typespec if you want to call
more attention to it than with the function VOID?.

So this finally bites the bullet and shifts to using ~null~ and ~void~.