Giant VOID! Implementation Shift
This commit contains the inclusion of void, and a move away from
the BLANK!-in-NULL-out concept to the VOID!-in-NULL-out idea.

It's kind of hard to do incrementally in a reasonable way, so it's
just brute force of changing a lot of things that are blanks to
being nulls... getting rid of the TRY incarnation that turned
nulls to blanks and replacing it with MAYBE, and just pushing that
through until things worked.

This is really just trying to sync up the bootstrap executable to
be useful.