Get rid of "OPT" native function
The OPT function was for turning blanks into nulls.  This idea is replaced
by REIFY which doesn't have anything to do with blanks in the modern
conception.  (In the modern conception, blanks do not have anything to
do with nulls, they are a placeholder, like the equivalent for arrays of
spaces in strings).