Make bootstrap possible in post-real-VOID world
This tweaks the rebmake process enough that it can continue building
with the old r3-8994d23, but also build with itself.

It required mental acrobatics to accomplish, and isn't particularly
pretty.  Unlike the idea in modern Ren-C where the old executable is
tweaked to the new rules, this tweaks the new executable backwards
to the old rules... so it continues using BLANK! in lots of places
for "nothingness" instead of trying to build mappings between
blank and null.

Only needs to be good enough to patch and maintain if for some reason
the bootstrap breaks and this has to be done again.