Get old %import-shim.r logic working in R3C
Given the bootstrap executable needs to change, it's a good time to
update that executable to do whatever low-hanging fruit makes it
more compatible with current practice.  This lets us get rid of the
circuitous concepts of mapping NULL => BLANK! and VOID! => NULL,
by having the bootstrap executable not error on plain null access
and also by distinguishing the void and trash types.

This significantly pares down the %bootstrap-shim.r, although the
bootstrap executable is not planning to add isotopes like splices...
so it's probably a more stable idea to simulate them using regular
blocks as the bootstrap has been doing.