Modernize TRANSCODE as much as possible
This changes TRANSCODE so that it is able to accept a string, in
which case it does an inefficient binary conversion and then for /NEXT
adds in the count of how many codepoints were consumed.

It can't go all-the-way-modern because that requires multi-returns,
so it passes in an ANY-WORD! of what variable to set with the rest of
the binary data in the /NEXT case.

The output result is always the transcoded data, not mixed in with the
"rest", which was always a problem.