Tweak `inline static` => INLINE or `static inline`
It doesn't really make a difference, but "best practices" favor you
writing `static inline` over `inline static`.

It also turns out that just inline is only visible from within a
translation unit unless it's explicitly marked extern, so the INLINE
macro (which in C++ resolves to inline) is as good as static inline.
So using the macro is fine.

This makes the change where it makes sense.  Also fixes an accidental
change to use the INLINE macro in libuv files.