Fix filename support in arrays
This old branch had buggy support for filenames put on arrays during
scan.  It was also trying to use Symbol* to do it, which was supposed
to cut down on duplication if you used DO or LOAD on the same file
more than once, but that's a drop in the bucket compared to the number
of series that you create from DO-ing and LOAD-ing in the first place.
And it meant there were Symbol* with illegal characters like spaces
and backslashes in them.

Having this be broken is particularly annoying when sorting out bootstrap
problems, so it was important to fix, even in this old executable.