Alternate multi-return strategy for SPLIT-PATH
This changes SPLIT-PATH to return the directory as a primary return
result and then set the filename via a variable passed as a refinement.

Doing it this way is closer to compatibility with the multi-return
version in modern Ren-C... in particular, if you only want the
directory back they are equivalent.