Resurrect BLANK! as reified "empty" alternative to []
While BLANK! has been removed from "falsey" or null-like character,
it makes sense to not be using null for places that want a light
surrogate for empty blocks, like the empty state of GLOM info.

This replaces it there, and restores blank behavior in series
enumerations as opting out... even though it is not a candidate
for "MAYBE"-ing to a void.