REPEAT => [FOR-NEXT COUNT-UP], LOOP => REPEAT
COUNT-UP will ultimately be subsumed by FOR.

    for x 3 [
       print [x]
    ]
    =>
    1
    2
    3

But some people like COUNT-UP better anyway, and FOR is still on the
back burner of design.

LOOP is being reserved for an ultra-powerful looping dialect which
has not been designed yet.  REPEAT is considered better for the case
of just giving a count and no variable to count on, anyway.