Kill split by tag, add split by void and quoted
There was no particular logic to the rationale of why the SPLIT
dialecting allowed you to pass in a TAG! and have it mean TAG!.  A
generic facility for splitting by things literally needs to be out of
band of the other argument types.

Future conversion of SPLIT to use a fluent UPARSE implementation is
pending the redevelopment of UPARSE for performance improvements.
This means splitting by splice isn't feasible at this time.

The bootstrap version of split--once thought needed to split by a
BAR!, is both no longer in use...and would have been doable with
a parse rule anyway (e.g. split by ['|]).