EVALfor BLOCK!, FRAME!, ERROR!, VARARGS!... not DO
DO and IMPORT are aiming to be polymorphic, which means you would be
able to DO a JavaScript file or IMPORT a CSSfile.

This means DO of TEXT! will no longer be supported, at least without
a header...because otherwise there's no way to know what language is
intended for the interpretation.

But additionally, DO of a BLOCK! would line up with aspects in the
IMPORT case of taking a block, to provide more parameters about the
fetching being done (like what version, etc.)

This moves to the long-awaited shift where EVAL is used instead of DO
for the more granular Rebol operations.