Eliminate HAS, enable MAKE OBJECT! in mezzanine
The R3C snapshot was taken at a weird time when MAKE OBJECT! did not
evaluate the block it received...it was all part of an attempt to
come up with a solution to having a "spec" and a "body" where the
spec could contain words not mentioned in the body, and any such
words would be known to be "non-valued".

That aspect has been solved with isotopes in a more clean way.  So
the object "spec" has been gotten rid of.  But this executable is
rigged up in a way that MAKE does not get a parent, so the only way
to get a parent in a MAKE operation is to use an arity-2 CONSTRUCT.

The model never really got pushed out, so there was a shim in
%mezz-legacy.r which translated MAKE PARENT [...] object calls into
CONSTRUCT calls.  Also, there was an arity-1 variant of the
evaluating CONSTRUCT that was given the name HAS.

Disrupting this too much would be an unnecessary investment (esp.
since it's not totally obvious that MAKE should work how it does,
and a distinct operator for extending objects is probably a good
idea).  However, the HAS operator means something new in modern
Ren-C for testing if words are in objects (IN is used for a more
general sense of binding into environments).

This moves the MAKE OBJECT! shim into the mezzanine so it can be
used instead of the arity-1 HAS, which is less confusing, and
reclaims HAS.