Eliminate INPUT in favor of ASK TEXT!
The concept in modern Ren-C is that INPUT is not a good name for a
function reading a line from the user, so the ASK is changed such that
if you pass it a datatype it will ask for that.  Hence ASK TEXT! is
equivalent to what INPUT was.

Also eliminates experimental old ECHO command (what it did is documented
on the forum):

https://forum.rebol.info/t/echo-off-e-g-for-now-no-echo-feature/1157