Shift away from capitalization for ASSERT_XXX
The assert() macro is not all uppercase, so this shifts to the
assumption that if you see Assert_XXX() in the name you know it's
something that's backed by a macro that vanishes in the release build.

Some other tweaks where routines that were once capitalized to point
at their "evil" nature are now checked at compile time that the evil
isn't having a bad effect, so the alarming all-caps names aren't
necessary anymore.