Series subclasses, in C++ only
An original attempt was to make it so there would be typechecking of
Series vs. Array, or Context, etc. in the C build, by making a struct
that contained the base struct as its only member.  This was a bit
questionable in terms of the standard and contaminated the source,
when an easier concept is to only have subclasses in the C++ build.

By adding subclasses, errors were caught where the UCS2 strings and
UTF-8 series were being mixed up.  This will also prevent mistakes
in the future if the bootstrap executable needs to be edited.