Eliminate REB_T_TRASH concept of "trashing"
There was a sort of intermediate "trashing" state of cells that didn't
have a very clear definition.  Modern Ren-C has not only taken the term
TRASH to mean antiform BLANK! as a central datatype (the contents of
unset variables), but it's also structured the ideas of cell states
to include "erased" and "poison", with an idea of "freshness" also.

This gets rid of the dead end REB_T_TRASH concept, and replaces its
usages with poison or erased cells.  The result should be better checked
(e.g. dropped stack cells would notice if they were being written to
instead of just read from, since poison blocks writes).