Make binary accessors require binary parameters
This was complicated a bit by the fact that BITSET! had cases of
doing VAL_BINARY_AT(), when it's not positional.  That was worked
around with asserts.

Modern Ren-C actually has a branch that doesn't use BINARY! for
bitsets at all, it's just fallen off the radar as a priority
to fix bitsets (which are completely broken in R3-Alpha).