Drop Fledgling Utf-8 Everywhere Scaffolding
The bootstrap build will never have UTF-8 Everywhere patched into
it, so there's no reason to keep the classes that were starting the
process toward UTF-8 Eerywhere around.

The syntax of Ucs2(*) and Ucs2_Next() etc. is kept just to give more
parity between the Utf8(*) in the modern implementation, in case any
snippets of code need to be ported back.