Rethink actions as having "targets" not "bindings"
At one point, all cells that used their "extra" slot as a node
phrased it in terms of being a "binding".  That has not stood the
test of time, as binding has come to mean something more specific
that is a property of words and arrays.

So this makes actions like definitional return which hold onto the
frame they want to jump to call those "targets".