Make parse flags match PARSE3 in current Ren-C
PARSE3 is a legacy R3-Alpha implementation of PARSE that is not the
current design state of the art (UPARSE).  But it is faster than that
state of the art--significantly so until UPARSE is restyled with
native code portions.

But it would be nice if R3C could do some modern constructs like
`pos: <here>` instead of just `pos:`, and used `<end>` instead of
`end`.  Then it would need a mode like current PARSE3 to run with
the "/REDBOL" conventions.

The current means for passing flags through PARSE3 is a mess, but
this goes ahead and synchronizes R3C with the code in PARSE3.
The reasoning for why all the flags were merged together was in
order to make it easier to run in stackless situations--since a
local `flags` variable would not persist.  That's not relevant
since it was decided that PARSE3 would not be implemented
stacklessly, but having the code match is useful.

By changing the flag mechanism, the /REDBOL flag can be passed
through.