Add handling of `<here>` tag to PARSE3/PARSE2
It's very confusing when reading PARSE code in bootstrap to see a lone
SET-WORD! given the new model.  Making it able to do `<here>` actually
moves the needle a bit on readability, enough so to make it possible
to do so when bootstrapping the mainline code.

This means PARSE/REDBOL does not support TAG! matching any longer (but
all Redbol compatibility is now sought through UPARSE, it is only for
bootstrapping that Redbol semantics are persisted at all--e.g. the
bootstrap of R3C)