Drop COLLECT/KEEP from PARSE3
General use of the feature never happened in PARSE3, partially because it
couldn't be used in bootstrap and not wanting to use something that was
potentially kind of flaky.

Now that the option is available to use it in bootstrap, it can't really
be because it can't be done the way it is in modern Ren-C without splices.
If it isn't used in bootstrap, then the potential maintenance cost of any
bugs in the feature outweighs its use.  The better decision seems to be
to scrap it from PARSE3.