Add support for SET-WORD! and ACROSS in PARSE3
If not using the /REDBOL mode, then SET-WORD! can be made to act in a
way that is a very poor imitation of the combinator capture of
synthesized results.  Making it so it's possible to do that can free
the bootstrap process for the new executable from being stuck with
the old syntax.