Change TRY in PARSE back to OPT, add OPTIONAL
For a time it was attempted to see if TRY should be the literate
substitute for the ugly abbreviation OPT in PARSE.  This was deemed
to be the wrong part of speech, and parallels drawn with Haskell's
PARSEC regarding the use of TRY there were incorrect.  This restores
OPT and gives OPTIONAL as a literate synonym for those who want it.