Disallow literal TAG! match in PARSE, other types
This reduces the number of legal types that can be matched literally in
blocks to just TEXT! and ISSUE!, since other types like TAG! are now
acting like combinators (to keep from saturating the space of WORD!
quite as much).

Consequently, some appearances of TAG! had to change to use THE in
order to be matched.