Replace PARSE ANY with OPT SOME
General thought on this is that there's no English word for "zero or
more", and the contention with better meanings for ANY is too great.
Refactorings have shown that being explicit about the OPT in OPT SOME
helps call out patterns and potential mistakes (e.g. where it's not
really OPT).