Use `<end>` tag instead of END word in bootstrap
This drops legacy support of the END word in PARSE3.

Using a TAG! instead of a WORD! has slightly more cost at the moment
in terms of detecting it, but it frees up the word for variables...which
is a tradeoff that was decided to make during the UPARSE design, then
ported back to PARSE3.