Remove progress requirement from SOME
Historically SOME would stop iterating if it detected progress was not
made.  A separate construct called WHILE did not have the requirement.

Modern UPARSE established that this was done with its own combinator
called FURTHER.  There does not appear to be a need for FURTHER in the
bootstrap process, so this just gets rid of it and deprecates WHILE when
the /REDBOL switch is not used.