Cache lib context index in symbol canon
There was a fledgling concept in older Ren-C builds that you could get to the
index for a symbol using the symbol itself as the place where multiple
binding indices were reachable.  This concept became "sea of words", where the
variables for named symbols in modules were a linked list coming from the
symbol itself.

That concept isn't going to be implemented in this old build.  But there was
an unused index in the stub which can be used to store the index of that
symbol in lib.  This commit implements storing that index, but doesn't yet
take advantage of it--instead verifying it in duplicate.