Take advantage of performance from previous commit
This leverages the performance improvements from the previous commit,
by not enumerating the lib context and adding its keys for binding...but
trusting the values in the canon symbol.

A significant amount of time is saved in running the bootstrap prep and
rebmake steps (about 10% empirically), so this was worth doing.