Beginnings of negated parser implementation
This starts the implementation of negatable parsers in UPARSE, to see
how viable the concept is.  It runs into a problem that there are datatype
combinators (like TAG! and WORD!) which need to propagate the /NEGATED
flag.  Only enough is implemented here to get negation on AHEAD and `<END>`
while the issue is thought about, and work on higher priority problems.