Make TRIM code match new executable
There was a note in the bootstrap-shim about TRIM that /AUTO needed
a patch:

    trim: adapt :trim [  ; there's a bug in TRIM/AUTO in 8994d23
        if auto [
            while [(not tail? series) and (series/1 = LF)] [
                take series
            ]
        ]
    ]

The semantics are not clear here, but /AUTO did indeed behave
differently in a test case.  The easiest thing to do is to make
the two implementations match, so this just brings the code from
the newer executable into the older one.