Get rid of SOMETHING? NOTHING? and TRASHIFY
The idea of a concept which could mean "either blank or null" was
considered interesting at some point and that was called "NOTHING?"

The isotopic world may have a similar need but it would likely be
that null or void would be two instances of nothing...yet it's slippery
enough that NULL-OR-VOID? is probably a better name for that concept
since you wouldn't know if it included trash, etc.

The only use of SOMETHING? was removed in the previous commit so this
goes ahead and gets rid of these natives.