Remove TRIM bootstrap shim, add tests
There was a claim that TRIM/AUTO needed a patch in the bootstrap shim.
It's not entirely clear what the problem was with that, but there was a
small difference in behavior.  To eliminate the difference, the current
code for TRIM was pasted into the bootstrap executable.

As part of looking into the behavior difference, this also incorporates
some trim tests from Red.