Improve Option(T) syntax, no parens needed
This uses a macro trick to make it possible to unwrap or try_unwrap
an Option(T) without needing to use parentheses:

Before:

    Option(T) opt_var = ...;
    if (opt_var) {
       T var = unwrap(var);
       ...
    }

After:

    Option(T) opt_var = ...;
    if (opt_var) {
       T var = unwrap var;
       ...
    }

It also defines `try_unwrap()` as just `maybe`

It's a subtle change, but reduces the noise enough that any function
which has optional output can use Option(T), vs. have a weird name
like `Maybe_Get_Whatever()`.  Because rather than encode the "Maybe"
in the function name and bypass use of the optional, you can
invoke it as `maybe Get_Whatever()`.

(Those changes coming in a subsequent commit.)