Rethink CONSTRUCT to not bind body to object
A new paradigm for member functions means it is not necessary
(nor even necessarily desirable) to have all references to ANY-WORD!
inside the spec block of an object construction routine to be
bound to that object's fields.

This leaves MAKE OBJECT! (a.k.a. what was known as CONTEXT) to do
the grand "everything in the block is bound" but CONSTRUCT as a
more modest operator that permits things like:

    make-point: func [x y] [
        return construct [x: x, y: y]
    ]