Fix stray binding code in rebTranscodeInto() API
TRANSCODE is not supposed to do any binding at all.  But some legacy code
was binding and using Get_Context_From_Stack() to do it...which was
bad in particular when writing code in the core that just randomly
started picking up contexts that should not apply.

This moves the specifier injection into the continuation pushing, which
is where it belonged.