Primitive implementation of :: as APPLY operator
The idea of using trailing slashes as a shorthand for invoking an
APPLY had brevity on its side, but it took that notation away from
better uses and wasn't that great.

Now that the :: SIGIL! is available, it is not a bindable WORD! and
hence has to have built-in evaluator behavior (if it does anything at
all).  It was picked as one of the first potentially interesting
symbols to use--almost being "too good" because :: might be useful
in the type system.  But APPLY is very important so, this goes ahead
and makes the shift.

The implementation is crude and only supports WORD! on the left hand
side for now, as well as a single unit of evaluation on the right.
But with plain APPLY as a fallback those shortcomings can be addressed
at a later date.