Make CALL default to returning trash, not 0
Since CALL will raise a definitional error if you don't use /RELAX if the
exit code of a process is not 0, the 0 is meaningless.

To help avoid people writing things like:

     if (1 = call [...]) [print "Exit code 1 signaled"]

This makes CALL without /RELAX return TRASH.  Trash is now not legal in
comparison operations, as UNSET! wasn't in Rebol2.