META requires /EXCEPT for raised errors
The casual META-ing of raised errors is rather dangerous, so this makes it
that you only get that if you explicitly use the /EXCEPT refinement.

(For the moment, caret ^ still has the semantics of META/EXCEPT)

This changes the behavior of PACK so that it will no longer pack up raised
errors.  Alternative solutions are applied to the one situation that
was used for:

  https://forum.rebol.info/t/no-raised-errors-in-pack-but-feel-the-power/2206