Simplify EVAL callsites based on NIHIL -> VOID Default
It seems that at some point, EVAL was made to automatically decay
NIHIL to VOID unless you specifically asked for the EVAL/UNDECAYED
option.  Not all clients consider a step that produces NIHIL to
be the same as one that produces VOID... in particular, if you
wanted to write your own stepwise EVAL that gave the same answer
as a non-stepwise one, you'd have to know the difference, or:

    >> my-eval-to-end [1 + 2 10 + 20 comment "hi"]
    == ~void~  ; anti (instead of 30)

But for clients who aren't trying to do that, getting VOID back
may well be just as good as NIHIL and a lot less trouble.

ASSERT was doing some rather convoluted things it didn't have to
be doing, as was the EMIT dialect in %prot-tls.r  This simplifies
those cases.