Switch EVAL/NEXT to Multi-Return, Add EVAL/ENTRAP
This implements the new design for EVAL/NEXT:

  https://forum.rebol.info/t/re-imagining-eval-next/767

As part of that design, the value returned from evaluation synthesis
appears in a parameter pack.  But raised errors cannot be put in PACKs.
However, if EVAL/NEXT propagated the raised error from the step as
its return value, then it wouldn't be able to provide the position
to continue from if that value was triaged.

  https://forum.rebol.info/t/should-evaluate-next-bomb-on-an-error/1491/3

This implements the proposal of EVAL/NEXT/ENTRAP which multiplexes
a raised error onto the value by META-representing all non-raised results,
and making plain ERROR! the result if it was raised.