Massive renaming: Series* -> Flex*
In the Rebol language itself, a series is a concept which puts together an
array pointer (allocated block of memory that holds cells) with an index.
Ren-C added a binding to that, since ANY-ARRAY! can carry a "specifier".

This means that a series is high level.  Yet inside the code, many places
referred to the array itself as a series.  (In R3-Alpha, these were REBSER
but still would declare variables like `REBSER series` or pass them to
functions like Append_To_Series())  This created a frustrating ambiguity
when going between the usermode experience and the internals, where the
term series in the internals was used for both instances of the value type
and the low-level array piece of that type.

This resolves that by changing the name of the lower level concept to
"Flex".  Besides being shorter than "Series" it helps realize that anywhere
in the code where a function or variable mentions series, that is talking
about a Cell with an index and a binding... not the resizable C vector
structure providing the memory and other aspects.

  https://forum.rebol.info/t/2221