Align bootstrap executable with TRANSCODE/NEXT return
The new convention is that the primary return result of TRANSCODE is
the remainder, with the transcoded element being the second item.

However, since it lacks multi-return the bootstrap executable has
to make the /NEXT refinement take a word to set for the transcoded
value.

This does that, but in order to avoid conflicts the bootstrap
executable calls its refinement /NEXT3.

(This seems like a good general strategy for the bootstrap executable
for such changes in convention.)