Drop transcode bootstrap forward shim
The bootstrap executable now has a /NEXT3 refinement that takes a WORD!
to write the transcoded value to.  This way it is not in contention
with the /NEXT refinement that doesn't take any args in the new Ren-C
(because it returns the value via multi-return).

The backwards shim then implements the /NEXT3 refinement.