Rename BINARY! to BLOB! (partial)
"Series" was confusingly both the low level construct, and the high-level
ANY-SERIES! with an index and binding... this was fixed by giving the
low-level construct a new name of "Flex".

A similar problem existed with "Binary" as the low level byte-bearing
allocation, and BINARY! having an index in it.  This takes a step to
solving that issue by making BLOB! the name of the exposed type that
is backed by a byte-sized Binary Flex.

(This only renames the internal types, the user-exposed type is still
BINARY! for the moment.  This is because it is a rewritten history
version from when the reverse was done--calling the store Blob and
leaving the user type as BINARY!.  That was deemed an inferior
choice to the reverse, so this commit is patched to fix it...the
actual renaming of BINARY!->BLOB! will be in a later commit.)