Make Get_Flex_Flag() and Get_Flex_Info() match master
The idea of using token pasting to put FLEX_FLAG_## into the macro for
testing flags helps avoid testing the wrong place (e.g. testing for info
flags with Get_Flex_Flag() or for leader flags with Get_Flex_Info())
It also makes the source briefer.

This goes ahead and makes that change for the bootstrap executable, and
actually caught an incorrect use testing for FLEX_INFO_INVALID on the
leader flags.